+++ packages/core/src/components/bal-button/bal-button.tsx
@@ -1,5 +1,7 @@
-import { Component, h, Prop, Host, Event, EventEmitter, ComponentInterface, Listen, Element } from '@stencil/core'
+import { Component, h, Prop, Host, Event, EventEmitter, ComponentInterface, Element, State } from '@stencil/core'
 import { Attributes, inheritAttributes } from '../../utils/attributes'
+import { rLCP } from '../../utils/helpers'
+import { ListenTo } from '../../utils/listen'
 
 @Component({
   tag: 'bal-button',
@@ -9,6 +11,8 @@ export class Button implements ComponentInterface {
   private inheritAttributes: Attributes = {}
 
   @Element() el!: HTMLElement
+
+  @State() isLargestContentPaintDone = false
 
   /**
    * The color to use from your application's color palette.
@@ -155,12 +159,16 @@ export class Button implements ComponentInterface {
    */
   @Event() balDidRender!: EventEmitter<BalEvents.BalButtonDidRenderDetail>
 
-  @Listen('click', { capture: true, target: 'document' })
+  @ListenTo('click', { capture: true, target: 'document' })
   listenOnClick(ev: UIEvent) {
     if (this.disabled && ev.target && ev.target === this.el) {
       ev.preventDefault()
       ev.stopPropagation()
     }
+  }
+
+  componentDidLoad(): void {
+    rLCP(() => (this.isLargestContentPaintDone = true))
   }
 
   componentWillLoad() {
@@ -310,15 +318,23 @@ export class Button implements ComponentInterface {
           data-testid="bal-button"
           {...ariaAttributes}
         >
-          <bal-spinner color={spinnerColor()} small {...this.loadingAttrs} deactivated={!this.loading} />
-          <bal-icon
-            {...this.leftIconAttrs}
-            class={this.square ? '' : 'icon-left'}
-            name={this.icon}
-            size={this.square ? this.size : 'small'}
-            turn={this.iconTurn}
-            inverted={this.isIconInverted}
-          />
+          {this.isLargestContentPaintDone && this.loading ? (
+            <bal-spinner color={spinnerColor()} small {...this.loadingAttrs} deactivated={!this.loading} />
+          ) : (
+            ''
+          )}
+          {this.isLargestContentPaintDone && this.icon ? (
+            <bal-icon
+              {...this.leftIconAttrs}
+              class={this.square ? '' : 'icon-left'}
+              name={this.icon}
+              size={this.square ? this.size : 'small'}
+              turn={this.iconTurn}
+              inverted={this.isIconInverted}
+            />
+          ) : (
+            ''
+          )}
           <span
             class={{
               'button-label': true,
@@ -330,14 +346,18 @@ export class Button implements ComponentInterface {
           >
             <slot />
           </span>
-          <bal-icon
-            {...this.leftRightAttrs}
-            class="icon-right"
-            name={this.iconRight}
-            size={'small'}
-            turn={this.iconTurn}
-            inverted={this.isIconInverted}
-          />
+          {this.isLargestContentPaintDone && this.iconRight ? (
+            <bal-icon
+              {...this.leftRightAttrs}
+              class="icon-right"
+              name={this.iconRight}
+              size={'small'}
+              turn={this.iconTurn}
+              inverted={this.isIconInverted}
+            />
+          ) : (
+            ''
+          )}
         </TagType>
       </Host>
     )