+++ packages/core/src/components/bal-carousel/bal-carousel.tsx
@@ -9,9 +9,8 @@ import {
   Event,
   EventEmitter,
   State,
-  Listen,
 } from '@stencil/core'
-import { raf } from '../../utils/helpers'
+import { debounce, raf, rLCP } from '../../utils/helpers'
 import { BEM } from '../../utils/bem'
 import { BalSlide, ControlItem } from './bal-carousel.type'
 import { TabControl } from './controls/tab-control'
@@ -28,6 +27,7 @@ import { getComputedWidth } from '../../utils/style'
 import { BalConfigState } from '../../interfaces'
 import { BalLanguage, ListenToConfig, defaultConfig } from '../../utils/config'
 import { i18nControlLabel } from './bal-carousel.i18n'
+import { ListenTo } from '../../utils/listen'
 
 @Component({
   tag: 'bal-carousel',
@@ -44,6 +44,7 @@ export class Carousel
   private carouselId = `bal-carousel-${CarouselIds++}`
   private carouselContainerId = `bal-carousel-${CarouselIds++}-container`
 
+  @State() isLargestContentfulPaintDone = false
   @State() isLastSlideVisible = true
   @State() isMobile = balBreakpoints.isMobile
   @State() language: BalLanguage = defaultConfig.language
@@ -132,12 +133,18 @@ export class Carousel
    * ------------------------------------------------------
    */
 
+  componentDidLoad(): void {
+    rLCP(() => {
+      this.isLargestContentfulPaintDone = true
+    })
+  }
+
   /**
    * LISTENERS
    * ------------------------------------------------------
    */
 
-  @Listen('touchmove', { target: 'window', passive: false })
+  @ListenTo('touchmove', { target: 'window', passive: false })
   async blockVerticalScrolling(ev: any) {
     if (!this.scrollY && this.el?.contains(ev.target)) {
       stopEventBubbling(ev)
@@ -171,7 +178,7 @@ export class Carousel
     this.itemsChanged()
   }
 
-  @Listen('keydown')
+  @ListenTo('keydown')
   listenToKeyDown(ev: KeyboardEvent) {
     if (this.htmlRole !== 'tablist') {
       if (ev.code === 'Tab') {
@@ -332,11 +339,13 @@ export class Carousel
     return undefined
   }
 
-  private async itemsChanged() {
-    const activeSlide = await this.buildSlide(this.value)
-
-    if (activeSlide) {
-      this.animate(activeSlide.transformActive, false)
+  private itemsChanged = debounce(() => this.itemsChangedInternal(), 100)
+  private async itemsChangedInternal() {
+    if (this.isLargestContentfulPaintDone) {
+      const activeSlide = await this.buildSlide(this.value)
+      if (activeSlide) {
+        this.animate(activeSlide.transformActive, false)
+      }
     }
   }
 
@@ -442,7 +451,7 @@ export class Carousel
           ...block.modifier(`controls-${this.controls}`).class(),
         }}
       >
-        {this.controls === 'tabs' ? (
+        {this.isLargestContentfulPaintDone && this.controls === 'tabs' ? (
           <TabControl
             value={this.value}
             items={controlItems}
@@ -492,7 +501,7 @@ export class Carousel
           </div>
         </div>
 
-        {this.controls === 'dots' ? (
+        {this.isLargestContentfulPaintDone && this.controls === 'dots' ? (
           <DotControl
             value={this.value}
             items={controlItems}
@@ -503,7 +512,7 @@ export class Carousel
           ''
         )}
 
-        {this.controls === 'large' ? (
+        {this.isLargestContentfulPaintDone && this.controls === 'large' ? (
           <LargeControl
             isFirst={this.isFirst()}
             isLast={this.isLast()}
@@ -519,7 +528,7 @@ export class Carousel
           ''
         )}
 
-        {this.controls === 'small' ? (
+        {this.isLargestContentfulPaintDone && this.controls === 'small' ? (
           <SmallControl
             isFirst={this.isFirst()}
             isLast={this.isLast()}