+++ packages/core/src/components/bal-dropdown/bal-dropdown.tsx
@@ -12,7 +12,7 @@ import {
   EventEmitter,
   Method,
 } from '@stencil/core'
-import { isArrowDownKey, isArrowUpKey, isEnterKey, isEscapeKey, isSpaceKey } from '@baloise/web-app-utils'
+import { isArrowDownKey, isArrowUpKey, isEnterKey, isEscapeKey, isSpaceKey } from '../../utils/keyboard'
 import { BEM } from '../../utils/bem'
 import { LogInstance, Loggable, Logger } from '../../utils/log'
 import { stopEventBubbling } from '../../utils/form-input'
@@ -43,7 +43,8 @@ import {
   defaultConfig,
 } from '../../utils/config'
 import { BalAriaForm, BalAriaFormLinking, defaultBalAriaForm } from '../../utils/form'
-import { waitAfterIdleCallback } from '../../utils/helpers'
+import { addEventListener, removeEventListener, waitAfterIdleCallback } from '../../utils/helpers'
+import { balBrowser } from '../../utils/browser'
 
 @Component({
   tag: 'bal-dropdown',
@@ -209,15 +210,38 @@ export class Dropdown
    * LIFECYCLE
    * ------------------------------------------------------
    */
-
+  hasConnected = false
   connectedCallback(): void {
-    this.eventsUtil.connectedCallback(this)
-    this.valueUtil.connectedCallback(this)
-    this.popupUtil.connectedCallback(this)
-    this.optionUtil.connectedCallback(this)
-    this.formSubmitUtil.connectedCallback(this)
-    this.focusUtil.connectedCallback(this)
-    this.autoFillUtil.connectedCallback(this)
+    if (!this.hasConnected) {
+      this.eventsUtil.connectedCallback(this)
+      this.valueUtil.connectedCallback(this)
+      this.popupUtil.connectedCallback(this)
+      this.optionUtil.connectedCallback(this)
+      this.formSubmitUtil.connectedCallback(this)
+      this.focusUtil.connectedCallback(this)
+      this.autoFillUtil.connectedCallback(this)
+    }
+
+    addEventListener(this.el, 'balOptionChange', this.listenToOptionChange)
+
+    if (balBrowser.hasDocument) {
+      addEventListener(document, 'click', this.listenOnClickOutside)
+      addEventListener(document, 'reset', this.resetHandler, {
+        capture: true,
+      })
+    }
+
+    this.hasConnected = true
+  }
+
+  disconnectedCallback(): void {
+    removeEventListener(this.el, 'balOptionChange', this.listenToOptionChange)
+    if (balBrowser.hasDocument) {
+      removeEventListener(document, 'click', this.listenOnClickOutside)
+      removeEventListener(document, 'reset', this.resetHandler, {
+        capture: true,
+      })
+    }
   }
 
   async componentWillRender() {
@@ -249,18 +273,15 @@ export class Dropdown
     this.httpFormSubmit = state.httpFormSubmit
   }
 
-  @Listen('balOptionChange')
-  async listenToOptionChange(ev: BalEvents.BalOptionChange) {
+  listenToOptionChange = (ev: BalEvents.BalOptionChange) => {
     this.optionUtil.listenToOptionChange(ev)
   }
 
-  @Listen('click', { target: 'document' })
-  listenOnClickOutside(ev: UIEvent) {
+  listenOnClickOutside = (ev: UIEvent) => {
     this.eventsUtil.handleOutsideClick(ev)
   }
 
-  @Listen('reset', { capture: true, target: 'document' })
-  resetHandler(ev: UIEvent) {
+  resetHandler = (ev: UIEvent) => {
     this.formSubmitUtil.handle(ev)
   }