+++ packages/core/src/components/bal-stage/bal-stage-image/bal-stage-image.tsx
@@ -1,7 +1,8 @@
-import { Component, h, ComponentInterface, Host, Prop, Element } from '@stencil/core'
+import { Component, h, ComponentInterface, Host, Prop, Element, State } from '@stencil/core'
 import { BEM } from '../../../utils/bem'
 import { Attributes } from '../../../interfaces'
 import { inheritAttributes } from '../../../utils/attributes'
+import { rLCP } from 'packages/core/src/utils/helpers'
 
 @Component({
   tag: 'bal-stage-image',
@@ -10,6 +11,7 @@ export class StageImage implements ComponentInterface {
   private imageInheritAttributes: Attributes = {}
 
   @Element() el!: HTMLElement
+  @State() isLargestContentfulPaintDone = false
 
   /**
    * set of images to be used as background image
@@ -21,22 +23,27 @@ export class StageImage implements ComponentInterface {
    */
   @Prop() fallback?: string
 
+  componentDidLoad(): void {
+    rLCP(() => {
+      this.isLargestContentfulPaintDone = true
+    })
+  }
+
   componentWillLoad() {
     this.imageInheritAttributes = inheritAttributes(this.el, ['alt'])
   }
 
   render() {
     const block = BEM.block('stage-image')
+    const imageSrc = this.fallback ? this.fallback : this.srcSet.split(',')[0]
 
     return (
       <Host class={{ ...block.class() }}>
-        <img
-          loading="lazy"
-          src={this.fallback ? this.fallback : this.srcSet.split(',')[0]}
-          srcset={this.srcSet}
-          sizes="100vw"
-          {...this.imageInheritAttributes}
-        />
+        {this.isLargestContentfulPaintDone ? (
+          <img src={imageSrc} srcset={this.srcSet} sizes="100vw" {...this.imageInheritAttributes} />
+        ) : (
+          ''
+        )}
       </Host>
     )
   }