+++ packages/core/src/components/bal-textarea/bal-textarea.tsx
@@ -9,7 +9,6 @@ import {
   Method,
   Watch,
   ComponentInterface,
-  Listen,
   State,
 } from '@stencil/core'
 import {
@@ -29,6 +28,7 @@ import { debounceEvent } from '../../utils/helpers'
 import { inheritAttributes } from '../../utils/attributes'
 import { BEM } from '../../utils/bem'
 import { BalAriaForm, BalAriaFormLinking, defaultBalAriaForm } from '../../utils/form'
+import { ListenTo } from '../../utils/listen'
 
 @Component({
   tag: 'bal-textarea',
@@ -169,14 +169,14 @@ export class Textarea implements ComponentInterface, FormInput<string | undefine
    */
   @Event() balFocus!: EventEmitter<BalEvents.BalTextareaFocusDetail>
 
-  @Listen('click', { capture: true, target: 'document' })
+  @ListenTo('click', { capture: true, target: 'document' })
   listenOnClick(ev: UIEvent) {
     inputListenOnClick(this, ev)
   }
 
   private resetHandlerTimer?: NodeJS.Timeout
 
-  @Listen('reset', { capture: true, target: 'document' })
+  @ListenTo('reset', { capture: true, target: 'document' })
   resetHandler(ev: UIEvent) {
     const formElement = ev.target as HTMLElement
     if (formElement?.contains(this.el)) {