+++ packages/core/src/utils/array.ts
@@ -1,0 +1,19 @@
+import isNil from 'lodash.isnil'
+import isEqual from 'lodash.isequal'
+
+/**
+ * Returns `true` if the arrays are equal
+ *
+ * ```typescript
+ * areArraysEqual(['a', 'b'], ['b', 'a']) // true
+ * ```
+ */
+export function areArraysEqual<T>(a: T[], b: T[]): boolean {
+  if (isNil(a) || isNil(b)) {
+    return false
+  }
+  const copyA = [...a].sort()
+  const copyB = [...b].sort()
+
+  return isEqual(copyA, copyB)
+}