+++ packages/core/src/utils/helpers.ts
@@ -26,6 +26,39 @@ import {
 declare const __zone_symbol__requestAnimationFrame: any
 declare const requestAnimationFrame: any
 
+/**
+ * Request Largest Contentful Paint (LCP) callback
+ */
+export const rLCP = (callback: () => void, timeout = 3000) => {
+  let isLargestContentPatinDone = false
+  if (!balBrowser.isSafari && balBrowser.hasWindow && 'PerformanceObserver' in window) {
+    const observer = new PerformanceObserver(entryList => {
+      const entries = entryList.getEntries()
+      const lcpEntry = entries[entries.length - 1] // Get the last (largest) entry
+      if (lcpEntry) {
+        // Disconnect the observer as we only need the LCP
+        observer.disconnect()
+
+        // Load the script after LCP
+        isLargestContentPatinDone = true
+        rIC(() => callback())
+      }
+    })
+
+    // Start observing for Largest Contentful Paint (LCP) entries
+    observer.observe({ type: 'largest-contentful-paint', buffered: true })
+
+    setTimeout(() => {
+      if (!isLargestContentPatinDone) {
+        observer.disconnect()
+        callback()
+      }
+    }, timeout)
+  } else {
+    return setTimeout(callback, 32)
+  }
+}
+
 export const rIC = (callback: () => void, timeout = 5000) => {
   if (balBrowser.hasWindow && 'requestIdleCallback' in window) {
     ;(window as any).requestIdleCallback(callback, { timeout })
@@ -46,6 +79,14 @@ export const debounceEvent = (ev: EventEmitter, wait: number): EventEmitter => {
     _original: ev,
     emit: debounce(original.emit.bind(original), wait),
   } as EventEmitter
+}
+
+export const debounceLCP = (func: (...args: any[]) => void, wait = 0) => {
+  let timer: any
+  return (...args: any[]): any => {
+    clearTimeout(timer)
+    timer = setTimeout(func, wait, ...args)
+  }
 }
 
 export const debounce = (func: (...args: any[]) => void, wait = 0) => {
@@ -281,8 +322,8 @@ export const waitForDesignSystem = async (el: any | null, _config?: BalConfig): 
       }),
     )
   }
-  await waitAfterFramePaint()
-  await waitAfterIdleCallback()
+
+  await Promise.all([waitAfterFramePaint(), waitAfterLargestContentfulPaintCallback(), waitAfterIdleCallback()])
 }
 
 export const waitAfterFramePaint = () => {
@@ -291,6 +332,10 @@ export const waitAfterFramePaint = () => {
 
 export const waitAfterIdleCallback = () => {
   return new Promise(resolve => rIC(() => runHighPrioritizedTask(resolve)))
+}
+
+export const waitAfterLargestContentfulPaintCallback = () => {
+  return new Promise(resolve => rLCP(() => runHighPrioritizedTask(resolve)))
 }
 
 export const runHighPrioritizedTask = (callback: (value: unknown) => void) => {