+++ packages/testing/src/commands/overrides/clear.command.ts
@@ -1,5 +1,4 @@
 import {
-  isDatepicker,
   isCheckbox,
   isInput,
   isRadio,
@@ -19,10 +18,6 @@ Cypress.Commands.overwrite<any, any>('clear', (originalFn: any, element: Cypress
 
   if (isCheckbox(element)) {
     return command(selectors.checkbox.input)
-  }
-
-  if (isDatepicker(element)) {
-    return command(selectors.datepicker.input)
   }
 
   if (isInput(element)) {